+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Fix Unifi Doorbell "Ding"
+
 ## 3.2.3-alpha.2 (2025-05-20)
 
 * (theimo1221) Add first version with Unifi-Access support